Avoid errors during first startup due to racy async behavior
The problem was that we didn't await refreshContextMenu() and then
called onConfigChange immediately. That caused
chrome.storage.onChanged.addListener() to try to
parse(changes.config.oldValue), which was not set on the very first
config change, when the config was first created.

Now we control the sequence of events and avoid race conditions.